requirements.txt
selenium
python-dotenv
webdriver-manager
pymongo[srv]
Flask